fix: Update Python SDK package name to slipkey-sdk
This commit changes the Python SDK package name from `slipkey-sdk-py` to `slipkey-sdk`.

Updates were made in:
- `sdk-py/setup.py`: Changed the `name` attribute in the `setup` function.
- `sdk-py/README.md`: Updated installation instructions to reflect the new package name.
