--- .github/workflows/ci.yml
@@ -1,43 +1,0 @@
-name: Slipkey SDK CI
-
-on:
-  push:
-    branches: [ "**" ] # Run on pushes to all branches
-  pull_request:
-    branches: [ "**" ] # Also run on pull requests to all branches
-
-jobs:
-  test-sdk:
-    runs-on: ubuntu-latest
-
-    strategy:
-      matrix:
-        node-version: [18.x, 20.x] # Test on current LTS and latest stable
-
-    steps:
-    - name: Checkout repository
-      uses: actions/checkout@v4
-
-    - name: Set up Node.js ${{ matrix.node-version }}
-      uses: actions/setup-node@v4
-      with:
-        node-version: ${{ matrix.node-version }}
-        cache: 'npm' # Cache npm dependencies
-        cache-dependency-path: 'slipkey-js/package-lock.json' # Path to package-lock
-
-    - name: Install dependencies for SDK
-      working-directory: ./slipkey-js
-      run: npm ci # Use ci for cleaner installs in CI
-
-    - name: Run tests for SDK
-      working-directory: ./slipkey-js
-      # The NODE_OPTIONS for ESM modules is already in the package.json test script.
-      # If it weren't, setting it here in `env` would be necessary:
-      # env:
-      #   NODE_OPTIONS: --experimental-vm-modules
-      run: npm test
-      # Note: The package.json script "test": "NODE_OPTIONS=--experimental-vm-modules ./node_modules/.bin/jest"
-      # already includes the necessary NODE_OPTIONS.
-      # If tests needed other env vars, they could be set here, e.g.:
-      # env:
-      #   TZ: 'UTC'
