--- .github/workflows/publish.yml
@@ -1,39 +1,0 @@
-name: Publish Slipkey SDK to NPM
-
-on:
-  push:
-    tags:
-      - 'v*.*.*' # Trigger on version tags like v1.0.0, v0.1.2, etc.
-
-jobs:
-  publish-npm:
-    runs-on: ubuntu-latest
-    # Important: Specify permissions if your repo requires them for package publishing
-    # permissions:
-    #   contents: read
-    #   packages: write # If publishing to GitHub Packages, or if needed for npm interaction
-
-    steps:
-    - name: Checkout repository
-      uses: actions/checkout@v4
-
-    - name: Set up Node.js
-      uses: actions/setup-node@v4
-      with:
-        node-version: '18.x' # Use an LTS version for publishing
-        registry-url: 'https://registry.npmjs.org/' # Specify npm registry
-        cache: 'npm'
-        cache-dependency-path: 'slipkey-js/package-lock.json' # Path to SDK's package-lock
-
-    - name: Install dependencies, build SDK
-      working-directory: ./slipkey-js
-      run: |
-        npm ci
-        npm run build       # For ESM type defs
-        npm run build:bundle # For UMD bundle (if part of package, ensure listed in "files")
-
-    - name: Publish to NPM
-      working-directory: ./slipkey-js
-      run: npm publish --access public
-      env:
-        NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
