--- examples/js/.gitignore
@@ -1,3 +1,0 @@
-node_modules/
-dist/
-package-lock.json
