--- examples/js/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "example",
-  "version": "1.0.0",
-  "type": "module",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "ts-node --esm src/example.ts"
-  },
-  "dependencies": {
-    "slipkey-sdk": "file:../../sdk-js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@types/node": "^22.15.29",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.8.3"
-  },
-  "description": ""
-}
