--- slipkey-js/jest.config.cjs
@@ -1,14 +1,0 @@
-module.exports = {
-  extensionsToTreatAsEsm: ['.ts'],
-  transform: {
-    '^.+\\.tsx?$': [
-      'ts-jest', // Use the string name, hoping local Jest resolves it
-      { useESM: true }
-    ]
-  },
-  moduleNameMapper: {
-    '^(\\.{1,2}/.*)\\.js$': '$1'
-  },
-  testEnvironment: 'node',
-  testMatch: ['**/tests/**/*.test.ts'],
-};
