--- slipkey-js/package.json
@@ -1,52 +1,0 @@
-{
-  "name": "slipkey-sdk",
-  "version": "0.1.0",
-  "description": "JavaScript/TypeScript SDK for the Slipkey protocol",
-  "type": "module",
-  "main": "dist/esm/index.js",
-  "exports": {
-    ".": "./dist/esm/index.js"
-  },
-  "scripts": {
-    "test": "NODE_OPTIONS=--experimental-vm-modules ./node_modules/.bin/jest",
-    "build": "tsc",
-    "build:bundle": "esbuild src/index.ts --bundle --outfile=dist/bundles/slipkey-sdk.umd.js --format=iife --global-name=SlipkeySDK"
-  },
-  "author": "",
-  "license": "MIT",
-  "keywords": [
-    "slipkey",
-    "proof-of-work",
-    "jwt",
-    "sdk",
-    "authentication",
-    "stateless"
-  ],
-  "homepage": "https://github.com/example/slipkey#readme",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/example/slipkey.git"
-  },
-  "bugs": {
-    "url": "https://github.com/example/slipkey/issues"
-  },
-  "types": "dist/esm/index.d.ts",
-  "files": [
-    "dist",
-    "README.md",
-    "LICENSE"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.5.14",
-    "esbuild": "^0.25.5",
-    "jest": "^29.7.0",
-    "ts-jest": "^29.3.4",
-    "typescript": "^5.8.3"
-  },
-  "dependencies": {
-    "jose": "^6.0.11"
-  }
-}
