--- slipkey-js/src/index.ts
@@ -1,9 +1,0 @@
-export { SlipkeyClient } from './client.js';
-export type { SlipkeyClientConfig } from './client.js';
-
-export { SlipkeyServer } from './server.js';
-export type { SlipkeyServerConfig, ServerCreditMetadata, CreditCalculationFunction } from './server.js';
-
-// It might also be useful to export some core crypto functions if they are intended for standalone use,
-// or types from 'jose' if users are expected to interact with them directly.
-// For now, focusing on the main client and server classes and their configs.
